CSHW-6

See merge request kingsbounty99/Homework!7

+++ src/Ozon.Route256.Practice.OrdersService/Repository/Dto/RegionDto.cs
@@ -3,6 +3,6 @@
 public sealed record RegionDto(
     long Id,
     string Name,
-    double Latitude,
-    double Longitude
+    decimal Latitude,
+    decimal Longitude
 );
